Move IjVimOptionGroup to correct package & rename
