+++ src/main/java/com/maddyhome/idea/vim/statistic/OptionsState.kt
@@ -21,7 +21,7 @@ import com.maddyhome.idea.vim.api.globalOptions
 import com.maddyhome.idea.vim.api.injector
 import com.maddyhome.idea.vim.options.OptionConstants
 import com.maddyhome.idea.vim.options.OptionScope
-import com.maddyhome.idea.vim.vimscript.services.IjOptionConstants
+import com.maddyhome.idea.vim.group.IjOptionConstants
 
 internal class OptionsState : ApplicationUsagesCollector() {
